Add PERFECT_MATCHING_DOUBLE definition to enforce doubles for PerfectMatching

+++ src/mwpm/stabiliser_graph.cpp
@@ -1,6 +1,6 @@
 #include "stabiliser_graph.h"
 
-int IStabiliserGraph::SpaceTimeDistance(int node1, int node2) const {
+double IStabiliserGraph::SpaceTimeDistance(int node1, int node2) const {
     int num_stab = GetNumStabilisers();
     if ((node1 < num_stab) && (node2 < num_stab)){
         return Distance(node1, node2);
