+++ src/mwpm/stabiliser_graph.h
@@ -4,8 +4,8 @@
 
 class IStabiliserGraph{
     public:
-        virtual int Distance(int node1, int node2) const = 0;
-        virtual int SpaceTimeDistance(int node1, int node2) const;
+        virtual double Distance(int node1, int node2) const = 0;
+        virtual double SpaceTimeDistance(int node1, int node2) const;
         virtual std::vector<int> ShortestPath(int node1, int node2) const = 0;
         virtual std::vector<int> SpaceTimeShortestPath(int node1, int node2) const;
         virtual int QubitID(int node1, int node2) const = 0;
